manage_users.yml: Add support for multiple inventories
So far, we've been assuming that a single inventory would be passed to this
playbook even though ansible allows us to use multiple inventories.
This patch adds support for multiple inventories by iterating through all
the inventory sources in order to find the cql_credentials section and then uses
the inventoy with this section to read the credentials for the internal users.