ansible-scylla-node: Make sure all nodes are up before adjusting replication
Before this patch, we were assuming that if we get to the adjust_keyspace_replication
task and start_scylla_service is set to true, it means that all the nodes are up,
since the 'start_scylla_service dependent tasks' block was already executed.
However, the role is responsible for starting scylla only for the nodes in
ansible_play_hosts, meaning that if the user has a cluster with 6 nodes, but is
running the role for 3 nodes (by using --limit), the role should start only these 3
nodes.
Having that in mind, this patch checks if ALL the nodes in the inventory (and not only
the ones in ansible_play_hosts) are up before adjusting the replication.

The adjust_keyspace_replication task was also adjusted to get the datacenter names of
all the nodes in the cluster and not only the ones in ansible_play_hosts.