ansible-scylla-node: Configure and start scylla only in the nodes in the 'play_hosts' var
Currently we're starting scylla service in all the nodes in the [scylla] section
in the inventory. By doing this we're not respecting the limits applied by the
user with the '--limit' option.
This patch fixes this by limiting the execution of the 'Start scylla non-seeds nodes serially'
task to the nodes in the 'play_hosts' var.
This patch does the same with the 'Create a map from dc to its list of nodes' task.