ansible-scylla-node: Validate if the role is being executed for the whole inventory
This patch takes advantage of the already existing full_inventory variable to
validate if the role is being executd for the whole inventory.
With the changes in this patch, by setting this variable to false, the user
should be able to run the role for a subset of the nodes in the scylla group.
This patch also replaces the usage of groups['scylla'] by ansible_play_hosts_all
in multiple places, since once we pass the validation of the full_inventory, we
know that:
  * We either are executing the role for the whole inventory, and in this case
    ansible_play_hosts_all == groups['scylla']
  * Or the user is explicitly setting full_inventory to false in order to run
    the role for a subset of nodes, and in this case ansible_play_hosts_all
    should be used in order to not touch other nodes in the group.

Note that if full_inventory is set to false, tasks that need to validate the state
of the whole cluster state like 'adjust_keyspace_replication', 'generate_tokens' and
'wait for the cluster to become healthy' will be skipped.