replace_node.yml: Move the RBNO check
In the next patch, an option will be added to conditionally skip the play 'Start Scylla in the new node',
which is also responsible for checking if RBNO is enabled.
Since the variable `_enable_rbno_grep` must always be defined (as it is used in other plays), the RBNO check needs to be moved up.
By moving the check up, we ensure `_enable_rbno_grep` is always set, and the play can then be skipped immediately afterward if required.