replace_node.yml: Skip 'Start Scylla in the new node' play if Scylla is already running
Currently, this playbook always try to find the replaced node in the storage_service API, and
this prevents the user from being able to retry the playbook if the replace already finished but
the playbook still failed for some reason (like a timeout, for example).
In order to allow the user to always retry the playbook, let's skip the play responsible for
querying the storage_service API and starting scylla service if the scylla-server service is already running.